website/requirements.txt
Django==4.2.6
asgiref==3.7.2
pytz==2022.1
sqlparse==0.4.4
python-decouple==3.8